+++ server/src/handlers/single_page_app.rs
@@ -1,9 +1,21 @@
-use crate::errors::BetterResult;
+use std::sync::Mutex;
+
+use crate::{appstate::AppState, errors::BetterResult};
 use actix_web::HttpResponse;
 
 /// Returns the atomic-data-browser single page application
-pub async fn single_page() -> BetterResult<HttpResponse> {
-    let body = include_str!("../../static/atomic-data-browser.html");
+pub async fn single_page(
+    data: actix_web::web::Data<Mutex<AppState>>,
+) -> BetterResult<HttpResponse> {
+    let context = data
+        .lock()
+        .expect("Failed to lock mutexguard in single_page");
+
+    let template = include_str!("../../static/atomic-data-browser.html");
+    let body = template
+        .replace("{script}", &context.config.script)
+        .replace("{asset_url}", &context.config.asset_url);
+
     let resp = HttpResponse::Ok()
         .content_type("text/html")
         // This prevents the browser from displaying the JSON response upon re-opening a closed tab